requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
colorama==0.4.6
    # via pytest
exceptiongroup==1.2.2
    # via pytest
iniconfig==2.0.0
    # via pytest
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
pytest==8.3.2
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
tomli==2.0.1
    # via pytest

# The following packages are considered to be unsafe in a requirements file:
# setuptools
